+++ src/BrickApp/Handle/Goto.hs
@@ -1,0 +1,34 @@
+-- | Event handler stuff for `GotoMode`.
+module BrickApp.Handle.Goto where
+
+import           Data.Maybe
+import           Control.Monad.IO.Class
+
+import qualified Brick.Main                    as M
+import qualified Brick.Types                   as T
+import           Brick.Widgets.Edit            as E
+import qualified Graphics.Vty                  as V
+import           Graphics.Vty.Input.Events      ( Event )
+
+import           BrickApp.Types
+import           BrickApp.Types.Names
+import           BrickApp.ModeAction.Goto
+
+gotoEventHandler
+  :: GopherBrowserState -> Event -> T.EventM AnyName (T.Next GopherBrowserState)
+gotoEventHandler gbs e = case e of
+    -- FIXME: esc quits! Change key...
+  V.EvKey V.KEsc   [] -> M.continue $ formerMode gbs
+  V.EvKey V.KEnter [] -> liftIO (mkGotoResponseState gbs) >>= M.continue
+  _                   -> M.continue =<< editorEventHandler gbs e
+ where
+  -- | A modification of the default Brick.Widgets.Edit event handler; changed to
+  -- return a GopherBrowserState instead of just an editor state.
+  editorEventHandler
+    :: GopherBrowserState -> Event -> T.EventM AnyName GopherBrowserState
+  -- TODO: e' is unused!
+  editorEventHandler _ e' =
+    -- Maybe this should be a general function in Representation.
+    let updateEditorInStatus x = gbs { gbsStatus = Just $ (fromJust $ gbsStatus gbs) { seEditorState = x } }
+    in  updateEditorInStatus
+          <$> E.handleEditorEvent e' (seEditorState $ fromJust $ gbsStatus gbs)